refactor: tx types
